+++ Example/Non-Card Payment Examples/Non-Card Payment Examples/BrowseExamplesViewController.m
@@ -47,7 +47,7 @@
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-    return 32;
+    return 33;
 }
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
@@ -148,6 +148,9 @@
             break;
         case 31:
             cell.textLabel.text = @"MobilePay";
+            break;
+        case 32:
+            cell.textLabel.text = @"Sunbit";
             break;
     }
     return cell;
@@ -351,6 +354,12 @@
         }
         case 31: {
             MobilePayExampleViewController *exampleVC = [MobilePayExampleViewController new];
+            exampleVC.delegate = self;
+            viewController = exampleVC;
+            break;
+        }
+        case 32: {
+            SunbitExampleViewController *exampleVC = [SunbitExampleViewController new];
             exampleVC.delegate = self;
             viewController = exampleVC;
             break;