+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 * [Added] When a card is saved (ie you're using a PaymentIntent + setup_future_usage or SetupIntent), legal disclaimer text now appears below the form indicating the card can be charged for future payments.
 * [Fixed] iOS 18 Compatibility with removing multiple saved payment methods
 * [Fixed] Fixed an issue where the keyboard could focus on a hidden phone number field.
+* [Added] Support for Sunbit (Private Beta) with PaymentIntents.
 
+### Payments
+* [Added] Support for Sunbit (Private Beta) bindings.
 
 ## 23.28.1 2024-07-16
 ### Payments